# activity/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Activity Service API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/activity-groups/{groupId}/leave": {
            "post": {
                "security": [
                    {
                        "BearerAuth": [],
                        "GatewayKeyAuth": []
                    }
                ],
                "description": "Allows the authenticated user to leave an activity group",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Members"
                ],
                "summary": "Leave an activity group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Activity Group ID",
                        "name": "groupId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully left the group",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
        "/api/v1/activity-groups/{groupId}/members": {
            "get": {
                "security": [
                    {
                        "BearerAuth": [],
                        "GatewayKeyAuth": []
                    }
                ],
                "description": "Lists members of an activity group with optional filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Members"
                ],
                "summary": "List members",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Activity Group ID",
                        "name": "groupId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (pending, confirmed, rejected, left, removed)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by role (creator, admin, member)",
                        "name": "role",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Limit results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset results",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of members",
                        "schema": {
                            "$ref": "#/definitions/api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_MemberListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": [],
                        "GatewayKeyAuth": []
                    }
                ],
                "description": "Invites a user to join an activity group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Members"
                ],
                "summary": "Invite a member to an activity group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Activity Group ID",
                        "name": "groupId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Invite data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.InviteMemberRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Member invited successfully",
                        "schema": {
                            "$ref": "#/definitions/api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_InviteMemberResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
        "/api/v1/activity-groups/{groupId}/members/{userId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": [],
                        "GatewayKeyAuth": []
                    }
                ],
                "description": "Retrieves a member by activity group and user ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Members"
                ],
                "summary": "Get a member",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Activity Group ID",
                        "name": "groupId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member data",
                        "schema": {
                            "$ref": "#/definitions/api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_MemberResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": [],
                        "GatewayKeyAuth": []
                    }
                ],
                "description": "Removes a member from an activity group",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Members"
                ],
                "summary": "Remove a member from an activity group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Activity Group ID",
                        "name": "groupId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member removed successfully",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
        "/api/v1/activity-groups/{groupId}/members/{userId}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": [],
                        "GatewayKeyAuth": []
                    }
                ],
                "description": "Approves a pending join request",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Members"
                ],
                "summary": "Approve a join request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Activity Group ID",
                        "name": "groupId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Join request approved",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "invalid_request"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string",
                    "example": "Invalid request"
                }
            }
        },
        "api.Response-any": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "only on success"
                },
                "error": {
                    "description": "only on error",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.APIError"
                        }
                    ]
                },
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                }
            }
        },
...
        "api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_MemberListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "only on success",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.MemberListResponse"
                        }
                    ]
                },
                "error": {
                    "description": "only on error",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.APIError"
                        }
                    ]
                },
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                }
            }
        },
...
        "github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.InviteMemberRequest": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "user_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426655440000"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "GatewayKeyAuth": {
            "description": "Key added automatically by the gateway when proxying requests; required directly here only when bypassing the gateway.",
            "type": "apiKey",
            "name": "X-Gateway-Key",
            "in": "header"
        }
    },
    "security": [
        {
            "GatewayKeyAuth": []
        }
    ]
}